+++ Editor/Core/SaintsPropertyDrawerUIToolkit.cs
@@ -700,27 +700,6 @@ namespace SaintsField.Editor.Core
 
         private static StyleSheet _noDecoratorDrawer;
 
-        private static bool PropertyIsDecoratorDrawer(PropertyAttribute propertyAttribute)
-        {
-            // ReSharper disable once ConvertIfStatementToReturnStatement
-            if (!_propertyAttributeToDecoratorDrawers.TryGetValue(propertyAttribute.GetType(),
-                    out IReadOnlyList<Type> eachDrawer))
-            {
-                // Debug.Log(propertyAttribute.GetType());
-                // foreach (Type key in PropertyAttributeToPropertyDrawers.Keys)
-                // {
-                //     if ($"{key}".Contains("SepTitle"))
-                //     {
-                //         Debug.Log(key);
-                //     }
-                // }
-                // not found
-                return false;
-            }
-
-            return eachDrawer.Any(drawerType => drawerType.IsSubclassOf(typeof(DecoratorDrawer)));
-        }
-
         private static VisualElement BindWatchUIToolkit(SerializedProperty property, Action<object> onValueChangedCallback, bool isReference, PropertyField propertyField, FieldInfo fieldInfo, object parent)
         {
             VisualElement trackerMain = propertyField.Q<VisualElement>(name: UIToolkitOnChangedTrackerName(property));